K-9 Mail notifications: Hopefully fix #7 and #8
K-9 Mail seems to completely ignore the where clause when querying for messages.
That is why we got data from the newest message instead of the one requested.
Work around by walking through the messages and match the uri manually.
This is not as expensive as I thought because in most cases it is the first
message is the one we want anyway.