Always send out the current state of GBDevice when asking for version
In a second step, request the version from the device (and send updated
values then)

RequestVersionInfo is either a misnomer or misused, depending on your view.
It is actually used by activities to get the current state of thde device.

We now provide this as quickly as possible, with the drawback of sometimes
sending results twice.