Autoconnect on startup if bt enabled
We should go to action connect on startup. This way if device reboots and bt is enabled and we have autoreconnect enabled the app will autoconnect to gadget without prompts.

On action connect if bt is disabled no use to show error. Try to enable bt with prompt, wait for adater to start and then continue connecting.