exclude not worn sleep times also from sleep sessions
This ensures that the sleep sessions times and the charts times are now
equal (correct behavior). At the same time this causes the sleep
sessions to be visually broken up into more small pieces and therefore
it now more highlights the issue of the band (mostly) wrongly marking
some samples as not worn.