Deprecate SMS Receiver
On recent android versions SMS can be handled through the normal notification flow.
Add "reply with SMS" functionality to the Call Control flow for refusing incoming call with SMS.
Also introduce a new preference screen for the deprecated functionalities.
Also add GBDeviceEventSendSMS device event to be used for outgoing SMS, not for replies.