Huami: Use activity summary timezone for details fetching (#3592)
It looks like some watches will reply with a fetch timestamp in a
different timezone than the request. When this happens, when we fetch
the activity details, the watch will reply with the details from the
wrong activity.

Use the same timezone in the activity details request as the watch
replies in the activity summary.